Add PPM reconstruction with symmetric van Albada limiter
- Fix VanAlbada equation to the correct symmetric form from the paper:
  f(a,b) = [a*(b^2+eps^2) + b*(a^2+eps^2)] / (a^2+b^2+2*eps^2)
- Merge ppm_van_albada.hpp into ppm_simple.hpp to eliminate code duplication
- Add ppm_va to Reconstruction enum
- Register ppm_va for hlle, hllc, and hlld Riemann solvers
- Add ppm_va to automated convergence test
